Valut API chnages after testing , ref #106

+++ src/angelleye/PayPal/rest/vault/CreditCardAPI.php
@@ -27,13 +27,18 @@ class CreditCardAPI {
             $requestArray = clone $creditCard;
             $creditCard->create($this->_api_context);     
             $returnArray=$creditCard->toArray();
+            $returnArray['REQUESTDATA']=$requestArray->toArray();
             $returnArray['RAWREQUEST']=$requestArray->toJSON();
             $returnArray['RAWRESPONSE']=$creditCard->toJSON();            
             return $returnArray;
         }
         catch (\PayPal\Exception\PayPalConnectionException $ex) {
-            return $ex->getData();        
-        }        
+            $errorReturnArray['ERRORS']=  json_decode($ex->getData(),true);
+            $errorReturnArray['REQUESTDATA']=$requestArray->toArray();
+            $errorReturnArray['RAWREQUEST']=$requestArray->toJSON();
+            $errorReturnArray['RAWRESPONSE']='';
+            return $errorReturnArray;
+        }
     }
         
     public function listAllCards($requestData) {   
@@ -43,31 +48,49 @@ class CreditCardAPI {
             $requestArray = json_encode($params);
             $cards = $creditCard->all($params, $this->_api_context);
             $returnArray=$cards->toArray();
+            $returnArray['REQUESTDATA']=$params;
             $returnArray['RAWREQUEST']=$requestArray;
-            $returnArray['RAWRESPONSE']=$cards->toJSON();            
-            return $returnArray;            
+            $returnArray['RAWRESPONSE']=$cards->toJSON();
+            return $returnArray;
         } catch (\PayPal\Exception\PayPalConnectionException  $ex) {
-            return $ex->getData();                
+            $errorReturnArray['ERRORS']=  json_decode($ex->getData(),true);
+            $errorReturnArray['REQUESTDATA']=$params;
+            $errorReturnArray['RAWREQUEST']=$requestArray;
+            $errorReturnArray['RAWRESPONSE']='';
+            return $errorReturnArray;
         }
     }
     
     public function showByID($requestData){        
         $creditCard = new \PayPal\Api\CreditCard();
             try {
-                $requestArray = clone $creditCard->setId($requestData['credit_card_id']);;
+                if(empty($requestData['credit_card_id'])){
+                    $errorReturnArray['ERRORS']= "Card ID is Empty.";                    
+                    return $errorReturnArray;
+                }
+                $requestArray = clone $creditCard->setId($requestData['credit_card_id']);
                 $card = $creditCard->get($requestData['credit_card_id'], $this->_api_context);
                 $returnArray=$card->toArray();
+                $returnArray['REQUESTDATA']=$requestArray->toArray();
                 $returnArray['RAWREQUEST']=$requestArray->toJSON();
                 $returnArray['RAWRESPONSE']=$card->toJSON();
                 return $returnArray;
             } catch (\PayPal\Exception\PayPalConnectionException  $ex) {
-                return $ex->getData();                
+                $errorReturnArray['ERRORS']=  json_decode($ex->getData(),true);
+                $errorReturnArray['REQUESTDATA']=$requestArray->toArray();
+                $errorReturnArray['RAWREQUEST']=$requestArray->toJSON();
+                $errorReturnArray['RAWRESPONSE']='';
+                return $errorReturnArray;
             }                    
     }
     
     public function deleteByID($requestData){        
         $creditCard = new \PayPal\Api\CreditCard();        
         try {
+            if(empty($requestData['credit_card_id'])){
+                $errorReturnArray['ERRORS']= "Card ID is Empty.";                    
+                return $errorReturnArray;
+            }
             $creditCard->setId($requestData['credit_card_id']);
             return $creditCard->delete($this->_api_context);                
         } catch (\PayPal\Exception\PayPalConnectionException  $ex) {
@@ -76,7 +99,11 @@ class CreditCardAPI {
     }
         
     public function UpdateCreditCard($requestData,$credit_card_id){
-            $creditCard = new \PayPal\Api\CreditCard();    
+            if(empty($credit_card_id)){
+                $errorReturnArray['ERRORS']= "Card ID is Empty.";                    
+                return $errorReturnArray;
+            }
+            $creditCard = new \PayPal\Api\CreditCard();
             $pathRequest = new \PayPal\Api\PatchRequest();
         try {
             $creditCard->setId($credit_card_id);
@@ -109,6 +136,7 @@ class CreditCardAPI {
                
                 $card = $creditCard->update($pathRequest,$this->_api_context);
                 $returnArray=$card->toArray();
+                $returnArray['REQUESTDATA']=$pathRequest->toArray();
                 $returnArray['RAWREQUEST']= $pathRequest->toJSON();
                 $returnArray['RAWRESPONSE']=$card->toJSON();
                 return $returnArray;                
@@ -117,7 +145,11 @@ class CreditCardAPI {
                 return "Fill Atleast One Array Field/Element";
             }
         } catch (\PayPal\Exception\PayPalConnectionException  $ex) {
-            return $ex->getData();                
+            $errorReturnArray['ERRORS']=  json_decode($ex->getData(),true);
+            $errorReturnArray['REQUESTDATA']=$pathRequest->toArray();
+            $errorReturnArray['RAWREQUEST']= $pathRequest->toJSON();
+            $errorReturnArray['RAWRESPONSE']='';
+            return $errorReturnArray;            
         }
     }
